Add a container class tying chat classes together
